+++ examples/react/src/main.tsx
@@ -16,14 +16,14 @@ import { stationWallet } from "@interchain-kit/station-extension";
 import { galaxyStationWallet } from "@interchain-kit/galaxy-station-extension";
 import { okxWallet } from "@interchain-kit/okx-extension";
 import { coin98Wallet } from "@interchain-kit/coin98-extension";
-import { Ledger, ledgerWallet } from "@interchain-kit/ledger";
+import { ledgerWallet } from "@interchain-kit/ledger";
 
 import { MockWallet } from "@interchain-kit/mock-wallet";
 import { starshipChain, starshipChain1 } from "./utils/starship.ts";
 import { ThemeProvider } from "@interchain-ui/react";
 
 const chainNames = [
-  "osmosistestnet",
+  // "osmosistestnet",
   "osmosis",
   "juno",
   "cosmoshub",
@@ -34,8 +34,6 @@ const chainNames = [
 // const chainNames = ["cosmoshub"];
 
 const walletConnect = new WCWallet();
-
-const ledgerConnect = new Ledger();
 
 const wallet1Mnemonic =
   "among machine material tide surround boy ramp nuclear body hover among address";
@@ -67,13 +65,13 @@ const mock2Wallet = new MockWallet(wallet2Mnemonic, _chains, {
 const _wallets: BaseWallet[] = [
   // mock1Wallet,
   // mock2Wallet,
-  // keplrWallet,
+  keplrWallet,
   // leapWallet,
   // cosmostationWallet,
   // stationWallet,
   // galaxyStationWallet,
   // walletConnect,
-  ledgerConnect,
+  ledgerWallet,
 ];
 
 ReactDOM.createRoot(document.getElementById("root")!).render(