+++ examples/vue/src/views/show.vue
@@ -1,0 +1,16 @@
+<script setup lang="ts">
+import { ref } from 'vue'
+import { useChain } from '@interchain-kit/vue';
+
+const chainName = ref('osmosistestnet')
+const { address } = useChain(chainName);
+</script>
+
+<template>
+  <div>
+    <div>address: {{ address }}</div>
+  </div>
+</template>
+
+<style scoped>
+</style>