+++ packages/react/src/hooks/useOfflineSigner.ts
@@ -1,4 +1,4 @@
-import { OfflineSigner } from '@interchainjs/cosmos/types/wallet';
+import { OfflineSigner, ICosmosGeneralOfflineSigner } from '@interchainjs/cosmos/types/wallet';
 import { useEffect, useState } from "react"
 import { useWalletManager } from "./useWalletManager"
 
@@ -6,7 +6,7 @@ export const useOfflineSigner = (chainName: string, walletName: string) => {
   const walletManager = useWalletManager()
   const wallet = walletManager.wallets.find((w) => w.info.name === walletName)
 
-  const [offlineSigner, setOfflineSigner] = useState<OfflineSigner | null>(null)
+  const [offlineSigner, setOfflineSigner] = useState<ICosmosGeneralOfflineSigner | null>(null)
 
   useEffect(() => {
     if (wallet && chainName) {