+++ packages/vue/src/index.ts
@@ -1,0 +1,5 @@
+
+
+export { default as ChainProvider } from './ChainProvider.vue';
+export * from './composables';
+export * from './modal';