+++ wallets/keplr-mobile/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.1-beta.38](https://github.com/@interchain-kit/keplr-mobile/compare/@interchain-kit/keplr-mobile@0.0.1-beta.37...@interchain-kit/keplr-mobile@0.0.1-beta.38) (2024-11-15)
+
+**Note:** Version bump only for package @interchain-kit/keplr-mobile
+
+## [0.0.1-beta.37](https://github.com/@interchain-kit/keplr-mobile/compare/@interchain-kit/keplr-mobile@0.0.1-beta.36...@interchain-kit/keplr-mobile@0.0.1-beta.37) (2024-11-11)
+
+**Note:** Version bump only for package @interchain-kit/keplr-mobile
+
+## [0.0.1-beta.36](https://github.com/@interchain-kit/keplr-mobile/compare/@interchain-kit/keplr-mobile@0.0.1-beta.35...@interchain-kit/keplr-mobile@0.0.1-beta.36) (2024-11-07)
+
+**Note:** Version bump only for package @interchain-kit/keplr-mobile
+
+## [0.0.1-beta.35](https://github.com/@interchain-kit/keplr-mobile/compare/@interchain-kit/keplr-mobile@0.0.1-beta.34...@interchain-kit/keplr-mobile@0.0.1-beta.35) (2024-11-04)
+
+**Note:** Version bump only for package @interchain-kit/keplr-mobile
+
+## [0.0.1-beta.34](https://github.com/@interchain-kit/keplr-mobile/compare/@interchain-kit/keplr-mobile@0.0.1-beta.33...@interchain-kit/keplr-mobile@0.0.1-beta.34) (2024-11-02)
+
+**Note:** Version bump only for package @interchain-kit/keplr-mobile
+
 ## [0.0.1-beta.33](https://github.com/@interchain-kit/keplr-mobile/compare/@interchain-kit/keplr-mobile@0.0.1-beta.32...@interchain-kit/keplr-mobile@0.0.1-beta.33) (2024-10-21)
 
 **Note:** Version bump only for package @interchain-kit/keplr-mobile